Introduce AbstractContextImpl and NinjaBootstrap to make it easier to create non-servlet containers.
